move jersey client builder
