+++ sda-commons-client-jersey/build.gradle
@@ -3,6 +3,7 @@ dependencies {
   api project(':sda-commons-server-opentracing')
   api project(':sda-commons-shared-tracing')
   api project(':sda-commons-shared-error')
+  api project(':sda-commons-server-opentelemetry')
 
   api 'io.dropwizard:dropwizard-client'
   api 'jakarta.servlet:jakarta.servlet-api'
@@ -21,6 +22,5 @@ dependencies {
   testImplementation 'org.assertj:assertj-core'
   testImplementation 'io.opentracing:opentracing-mock'
 
-  implementation 'io.opentelemetry.instrumentation:opentelemetry-apache-httpclient-4.3'
   testImplementation 'io.opentelemetry:opentelemetry-sdk-testing'
 }
